refactor: 패키지 분리 및 의존성 사이클 제거
