+++ src/main/java/controller/MainController.java
@@ -1,11 +1,16 @@
 package controller;
 
-import annotation.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.FileIoUtils;
+import was.annotation.Controller;
+import was.annotation.Mapping;
+import was.annotation.RequestMethod;
+import was.domain.response.Response;
+import was.domain.response.StatusCode;
+import was.domain.response.Version;
+import was.utils.FileIoUtils;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
