+++ src/main/java/was/annotation/Controller.java
@@ -1,11 +1,11 @@
-package annotation;
+package was.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface QueryString {
+public @interface Controller {
 }
