+++ src/main/java/was/utils/ParamsParser.java
@@ -1,4 +1,4 @@
-package utils;
+package was.utils;
 
 import lombok.Getter;
