+++ src/main/java/was/utils/PathParamsParser.java
@@ -1,4 +1,4 @@
-package utils;
+package was.utils;
 
 import lombok.Getter;
