+++ src/main/java/webserver/WebServer.java
@@ -2,6 +2,7 @@ package webserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import was.handler.RequestHandler;
 
 import java.net.ServerSocket;
 import java.net.Socket;
@@ -11,7 +12,6 @@ public class WebServer {
     private static final int DEFAULT_PORT = 8080;
 
     public static void main(String args[]) throws Exception {
-//        System.out.println(Class.forName("Controller"));
         int port = 0;
         if (args == null || args.length == 0) {
             port = DEFAULT_PORT;
