+++ src/test/java/webserver/RequestHandlerTest.java
@@ -2,7 +2,8 @@ package webserver;
 
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
-import utils.FileIoUtils;
+import was.handler.RequestHandler;
+import was.utils.FileIoUtils;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
